requirements.txt
streamlit
tensorflow
Pillow
numpy
gdown
pillow-heif